Refactors SecurityExceptionHandler to PasswordResetHandler
Renames and narrows the scope of the security exception handler to focus specifically on password reset functionality, improving separation of concerns in the security module.

The change reflects a more focused responsibility for handling password reset-related exceptions, making the codebase more maintainable and modular.
