+++ src/main/java/info/mackiewicz/bankapp/security/exception/handler/ExceptionToErrorMapper.java
@@ -1,0 +1,9 @@
+package info.mackiewicz.bankapp.security.exception.handler;
+
+import info.mackiewicz.bankapp.shared.exception.handlers.ErrorCode;
+
+public interface ExceptionToErrorMapper {
+
+    ErrorCode map(Exception ex);
+
+}
