+++ src/main/java/info/mackiewicz/bankapp/security/exception/handler/RequestUriHandler.java
@@ -1,0 +1,13 @@
+package info.mackiewicz.bankapp.security.exception.handler;
+
+import org.springframework.stereotype.Component;
+import org.springframework.web.context.request.ServletWebRequest;
+import org.springframework.web.context.request.WebRequest;
+
+@Component
+public class RequestUriHandler {
+
+        public String getRequestURI(WebRequest request) {
+        return ((ServletWebRequest) request).getRequest().getRequestURI();
+    }
+}
